+++ CountdownApp/Components/Fonts/FontSize.swift
@@ -1,0 +1,15 @@
+import UIKit
+
+enum FontSize: CGFloat, CaseIterable {
+    case xSmall = 10
+    case small = 12
+    case medium = 14
+    case large = 16
+    case xLarge = 18
+    case xxLarge = 20
+    case xxxLarge = 24
+    case huge = 32
+    case giant = 40
+    case xGiant = 48
+    case xxGiant = 56
+}