+++ CountdownApp/Components/Fonts/FontWeight.swift
@@ -1,0 +1,7 @@
+public enum FontWeight: String, CaseIterable {
+    case light = "Light"
+    case regular = "Regular"
+    case medium = "Medium"
+    case semibold = "SemiBold"
+    case bold = "Bold"
+}