+++ changelog.md
@@ -3,3 +3,7 @@
 ### All notable changes to CountdownApp should be document in this file
 
 - [COUN-9]: Config: Remove StoryBoard from Project
+- [N/A]: Feature: Add Class Auto Layout
+- [N/A]: Feature: Add Class Assets
+- [N/A]: Feature: Add Fonts
+- [N/A]: